requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
pydantic>=2.0.0
requests
